requirements.txt
Flask==2.3.3
requests==2.31.0
PyJWT[crypto]==2.8.0
cryptography==41.0.7
gunicorn==21.2.0